:truck: Rename internal module to `pyaigverse` for clearer separation of concerns
