+++ bindings/aigverse/CMakeLists.txt
@@ -1,19 +1,19 @@
-pybind11_add_module(aigverse
+pybind11_add_module(pyaigverse
         # Prefer thin LTO if available
         THIN_LTO
         aigverse.cpp)
-target_link_libraries(aigverse PRIVATE mockturtle)
-target_include_directories(aigverse
+target_link_libraries(pyaigverse PRIVATE mockturtle)
+target_include_directories(pyaigverse
         PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/include)
-set_property(TARGET aigverse PROPERTY POSITION_INDEPENDENT_CODE ON)
+set_property(TARGET pyaigverse PROPERTY POSITION_INDEPENDENT_CODE ON)
 
 if (MSVC)
-    target_compile_options(aigverse PRIVATE /utf-8)
+    target_compile_options(pyaigverse PRIVATE /utf-8)
     add_definitions(-D_WIN32_WINNT=0x0A00)
     add_definitions(-DUNICODE -D_UNICODE)
 endif()
 
 # Install directive for scikit-build-core
-install(TARGETS aigverse
+install(TARGETS pyaigverse
         DESTINATION .
         COMPONENT aigverse_Python)
